requirements.txt
cffi==1.2.1
cryptography==1.0.2
Django==1.8.5
djangorestframework==3.2.4
enum34==1.0.4
httpie==0.9.2
idna==2.0
ipaddress==1.0.14
ndg-httpsclient==0.4.0
psycopg2==2.6.1
pyasn1==0.1.9
pycparser==2.14
Pygments==2.0.2
pyOpenSSL==0.15.1
python-magic==0.4.10
requests==2.8.0
six==1.10.0
uWSGI==2.0.11.2
wheel==0.24.0
